refactor(consumption): extract MAF-deviation helpers out of maintenance_analyzer (Refs #563 phase: maintenance_analyzer)
Brings maintenance_analyzer.dart under the 300-LOC budget by moving
the MAF-deviation heuristic (`_detectMafDeviation` + `_medianCruiseFuelRate`)
into a sibling part file `_maf_deviation_detector.dart`. Public API
(`analyzeMaintenance`, `MaintenanceAnalyzerThresholds`) is unchanged;
library-private helpers (`_TimedValue`, `_emitHalfSplitSignal`, `_median`)
stay private via `part`/`part of`.

LOC: 322 -> 269 (analyzer) + 61 (new sibling). Existing
maintenance_analyzer_test.dart (10 cases) passes unchanged — no behavior
change.
